+++ tomni/annotation_manager/annotations/polygon/main.py
@@ -1,17 +1,24 @@
 import warnings
 from dataclasses import asdict
-from typing import List
+from typing import List, Type
 
 import cv2
 import numpy as np
 
-from ...utils import parse_points_to_contour
+from ...utils import parse_points_to_contour, are_lines_equal, simplify_line
 from ..annotation import Annotation
 from ..point import Point
 
 
 class Polygon(Annotation):
-    def __init__(self, points: List[Point], id, label, children, parents):
+    def __init__(
+        self,
+        points: List[Point],
+        id: str,
+        label: str = "",
+        children: List[Annotation] = [],
+        parents: List[Annotation] = [],
+    ):
         """Initializes a Polygon object.
 
         Args:
@@ -24,7 +31,7 @@ class Polygon(Annotation):
         MIN_NR_POINTS = 3
 
         super().__init__(id, label, children, parents)
-        self._points: List[Point] = points
+        self._points: List[Point] = simplify_line(points)
         self._contour: List[np.ndarray] = parse_points_to_contour(points)
 
         self._has_enough_points = len(points) >= MIN_NR_POINTS
@@ -67,7 +74,7 @@ class Polygon(Annotation):
 
     @property
     def convex_hull_area(self) -> float:
-        """Convex Hull Area by cv2 contour operations. 
+        """Convex Hull Area by cv2 contour operations.
 
         Returns:
             float: Polygon's convex hull area.
@@ -101,6 +108,10 @@ class Polygon(Annotation):
     def points(self) -> List[Point]:
         return self._points
 
+    @points.setter
+    def points(self, *arg, **kwargs) -> None:
+        raise SyntaxError("Points are Immutable")
+
     @property
     def roundness(self) -> float:
         """Roundness: Area / (radius_enclosing_circle**2 * pi).
@@ -116,7 +127,7 @@ class Polygon(Annotation):
     def to_dict(self, decimals: int = 2) -> dict:
         polygon_dict = {
             "type": "polygon",
-            "points": [asdict(point) for point in self._points],
+            "points": [asdict(point) for point in self.points],
         }
 
         if self._has_enough_points:
@@ -148,7 +159,7 @@ class Polygon(Annotation):
         if not self._perimeter:
             self._calculate_perimeter()
 
-        self._circularity = (4 * np.pi * self._area) / (self._perimeter ** 2)
+        self._circularity = (4 * np.pi * self._area) / (self._perimeter**2)
 
     def _calculate_convex_hull_area(self) -> None:
         if not self._has_enough_points:
@@ -168,5 +179,18 @@ class Polygon(Annotation):
             self._calculate_area()
 
         _, radius = cv2.minEnclosingCircle(self._contour)
-        enclosing_circle_area = radius ** 2 * np.pi
+        enclosing_circle_area = radius**2 * np.pi
         self._roundness = self._area / enclosing_circle_area
+
+    def __eq__(self, other):
+        if not isinstance(other, Polygon):
+            # don't attempt to compare against unrelated types
+            return False
+
+        are_points_equal = are_lines_equal(self.points, other.points, is_enclosed=True)
+        reverse_points = other.points
+        reverse_points.reverse()
+        are_points_equal_mirrored = are_lines_equal(
+            self.points, reverse_points, is_enclosed=True
+        )
+        return are_points_equal | are_points_equal_mirrored