+++ tomni/annotation_manager/annotations/polygon/test_polygon_comparing.py
@@ -1,0 +1,92 @@
+from unittest import TestCase
+
+from tomni.annotation_manager import Point, Polygon
+
+
+class TestPolygonCompare(TestCase):
+    def test_positive_squares(self):
+        poly1 = Polygon(
+            [Point(1, 1), Point(1, 2), Point(2, 2), Point(2, 1)], id=1, label="square"
+        )
+        poly2 = Polygon(
+            [Point(1, 1), Point(1, 2), Point(2, 2), Point(2, 1)], id=2, label="square"
+        )
+
+        self.assertTrue(poly1 == poly2)
+
+    def test_flipped_squares(self):
+        poly1 = Polygon(
+            [Point(1, 1), Point(1, 2), Point(2, 2), Point(2, 1)], id=1, label="square"
+        )
+        poly2 = Polygon(
+            [Point(2, 2), Point(2, 1), Point(1, 1), Point(1, 2)],
+            id=2,
+            label="flipped square",
+        )
+
+        self.assertTrue(poly1 == poly2)
+
+    def test_rotated_squares(self):
+        poly1 = Polygon(
+            [Point(1, 1), Point(1, 2), Point(2, 2), Point(2, 1)], id=1, label="square"
+        )
+        poly2 = Polygon(
+            [Point(1, 1), Point(2, 1), Point(2, 2), Point(1, 2)],
+            id=2,
+            label="rotated square",
+        )
+
+        self.assertTrue(poly1 == poly2)
+
+    def test_rotated_and_flipped_squares(self):
+        poly1 = Polygon(
+            [Point(1, 1), Point(1, 2), Point(2, 2), Point(2, 1)], id=1, label="square"
+        )
+        poly2 = Polygon(
+            [Point(1, 1), Point(2, 1), Point(2, 2), Point(1, 2)],
+            id=2,
+            label="flipped rotated square",
+        )
+
+        self.assertTrue(poly1 == poly2)
+
+    def test_positive_squares_extra_points(self):
+        poly1 = Polygon(
+            [Point(1, 1), Point(1, 30), Point(30, 30), Point(30, 1)],
+            id=1,
+            label="square",
+        )
+        poly2 = Polygon(
+            [
+                Point(1, 1),
+                Point(1, 30),
+                Point(15, 30),
+                Point(30, 30),
+                Point(30, 14),
+                Point(30, 1),
+            ],
+            id=2,
+            label="square",
+        )
+
+        self.assertTrue(poly1 == poly2)
+
+    def test_positive_squares_hourglass(self):
+        poly1 = Polygon(
+            [Point(1, 1), Point(1, 2), Point(2, 2), Point(2, 1)], id=1, label="square"
+        )
+        poly2 = Polygon(
+            [Point(1, 1), Point(2, 2), Point(1, 2), Point(2, 1)],
+            id=2,
+            label="hourglass",
+        )
+
+        self.assertFalse(poly1 == poly2)
+
+    def test_positive_wrong_type(self):
+        poly1 = Polygon(
+            [Point(1, 1), Point(1, 2), Point(2, 2), Point(2, 1)], id=1, label="square"
+        )
+        point2 = Point(1, 1)
+
+        self.assertFalse(poly1 == point2)