+++ HISTORY.rst
@@ -12,9 +12,12 @@
 - CDF-Main: Implement `__add__` and `__radd__`.
 - CDF-Main: Implement `append`
 - Add polygon annotation class.
+- CDF-polygon: Implement `__eq__`
+- CDF-polygon: Remove useless point
 - Add ellipse annotation class.
 - CDF-ellipse: Implement `__eq__`
 - CDF-ellipse: Set all rotations between 0 and 90, flip radii if needed
+- Renamed `AnnotationManager` to `AnnotationManager`
 
 1.15.0 (2022-12-05)
 ------------------